[Core] Rename logging module to log
The logging module was shadowing the built-in Python module of the same
name, which could cause confusion and hard-to-debug bugs.

Closes #58
