+++ doc/src/ForHosts.dox
@@ -84,10 +84,10 @@
  *    - An instance of the custom class derived from
  *      @ref openassetio.hostAPI.HostInterface "HostInterface".
  *    - A `logger` (derived from
- *      @ref openassetio.logging.LoggerInterface "LoggerInterface"), for
+ *      @ref openassetio.log.LoggerInterface "LoggerInterface"), for
  *      simple console logging you can use a
- *      @ref openassetio.logging.ConsoleLogger "ConsoleLogger" wrapped
- *      in a @ref openassetio.logging.SeverityFilter "SeverityFilter".
+ *      @ref openassetio.log.ConsoleLogger "ConsoleLogger" wrapped
+ *      in a @ref openassetio.log.SeverityFilter "SeverityFilter".
  *    - A `managerFactory` capable of instantiating managers,
  *      in the majority of cases a default-configured @ref
  *      openassetio.pluginSystem.PluginSystemManagerFactory
